+++ B6.0/Plugins/src/Video5.10/DaedalusGraphics.cpp
@@ -74,6 +74,7 @@ extern bool g_bUseSetTextureMem;
 extern DWORD g_maxTextureMemUsage;
 
 #include "../../../config.h"
+extern int loadbForceClearTextures();
 extern bool bloadstate[MAX_SAVE_STATES];
 extern bool bsavestate[MAX_SAVE_STATES];
 extern bool bload1964state[MAX_SAVE_STATES];
@@ -625,8 +626,11 @@ void _VIDEO_RICE_510_UpdateScreen (void)
 			if (bloadstate[i]) // This will never be true, unless UHLE accepts plugins. 
 			{
 				try{
-					bPurgeOldBeforeIGM = TRUE;
-					gTextureCache.PurgeOldTextures();
+					if (loadbForceClearTextures())
+					{
+						bPurgeOldBeforeIGM = TRUE;
+						gTextureCache.PurgeOldTextures();
+					}
 					__EMU_LoadState(i);
 				}catch(...){};
 
@@ -638,8 +642,11 @@ void _VIDEO_RICE_510_UpdateScreen (void)
 			else if (bload1964state[i]) 
 			{
 				try{
-					bPurgeOldBeforeIGM = TRUE;
-					gTextureCache.PurgeOldTextures();
+					if (loadbForceClearTextures())
+					{
+						bPurgeOldBeforeIGM = TRUE;
+						gTextureCache.PurgeOldTextures();
+					}
 					__EMU_LoadState(i);
 				}catch(...){};
 
@@ -651,8 +658,11 @@ void _VIDEO_RICE_510_UpdateScreen (void)
 			else if (bloadPJ64state[i]) 
 			{
 				try{
-					bPurgeOldBeforeIGM = TRUE;
-					gTextureCache.PurgeOldTextures();
+					if (loadbForceClearTextures())
+					{
+						bPurgeOldBeforeIGM = TRUE;
+						gTextureCache.PurgeOldTextures();
+					}
 					__EMU_LoadState(i);
 				}catch(...){};
